Refactor closeIcon
We are allowing icons to be a component or an image.
For the close button icon we forget to correctly render the component.
Also we have split the `closeIcon` into two different icons `settingsCloseIcon` and `popUpCloseIcon`, same with the className `closeButtonClassName` into `settingsCloseButtonClassName` and `popUpCloseButtonClassName` so we can easily manage settings close button and popup close button with different classNames and icons.